// deploy_ios/src/app/models/userdetails.ts
export class UserDetails {
    id: string = "";
    username : string = "";
    password: string = "";
    email: string = "";
    creditCardNumber: string = "";
    expiry: string = "";
    addressLine1: string = "";
    addressLine2: string = "";
    city: string = "";
    postalCode: string = "";
    countryCode: string = "";
    key1: string = "";
    key2: string = "";
}